Move coll-ik solve to solve_ik
