+++ examples/03_kin_with_coll.py
@@ -6,22 +6,20 @@ from typing import Optional, Literal
 from pathlib import Path
 import time
 import jax
-import jaxls
 
 from loguru import logger
 import tyro
 
 import jax.numpy as jnp
-import jax_dataclasses as jdc
 import jaxlie
 import numpy as onp
 
 import viser
 import viser.extras
 
-from jaxmp import JaxKinTree, RobotFactors
-from jaxmp.coll import Plane, RobotColl, Sphere, CollGeom, link_to_spheres
-from jaxmp.extras.urdf_loader import load_urdf
+from jaxmp import JaxKinTree
+from jaxmp.coll import Plane, RobotColl, CollGeom, link_to_spheres, Sphere
+from jaxmp.extras import load_urdf, solve_ik_with_coll
 
 
 def main(
@@ -208,98 +206,6 @@ def main(
                 _robot_coll.world_coll_dist(kin, joints, ground_obs).item(),
                 _robot_coll.world_coll_dist(kin, joints, curr_sphere_obs).item(),
             )
-
-
-@jdc.jit
-def solve_ik_with_coll(
-    kin: JaxKinTree,
-    target_joint_indices: jnp.ndarray,
-    target_poses: jaxlie.SE3,
-    robot_coll: RobotColl,
-    world_coll: list[CollGeom],
-    initial_pose: jnp.ndarray,
-    *,
-    pos_weight: float = 5.0,
-    rot_weight: float = 1.0,
-    rest_weight: float = 0.001,
-    limit_weight: float = 100.0,
-    self_coll_weight: float = 5.0,
-    world_coll_weight: float = 10.0,
-) -> jnp.ndarray:
-    # Create factor graph.
-    factors: list[jaxls.Factor] = []
-
-    JointVar = RobotFactors.get_var_class(kin, initial_pose)
-    joint_var_idx = 0
-
-    ik_weight = jnp.array([pos_weight] * 3 + [rot_weight] * 3)
-    factors.extend(
-        [
-            RobotFactors.ik_cost_factor(
-                JointVar,
-                joint_var_idx,
-                kin,
-                target_poses,
-                target_joint_indices,
-                ik_weight,
-            ),
-            RobotFactors.rest_cost_factor(
-                JointVar,
-                joint_var_idx,
-                jnp.array([rest_weight] * kin.num_actuated_joints),
-            ),
-            RobotFactors.limit_vel_cost_factor(
-                JointVar,
-                joint_var_idx,
-                kin,
-                0.01,
-                jnp.array([limit_weight] * kin.num_actuated_joints),
-                initial_pose,
-            ),
-            RobotFactors.limit_cost_factor(
-                JointVar,
-                joint_var_idx,
-                kin,
-                jnp.array([limit_weight] * kin.num_actuated_joints),
-            ),
-        ]
-    )
-
-    # Add collision factors.
-    self_coll_factor = RobotFactors.self_coll_factor(
-        JointVar, joint_var_idx, kin, robot_coll, 0.05, self_coll_weight
-    )
-    world_coll_factors = [
-        RobotFactors.world_coll_factor(
-            JointVar, joint_var_idx, kin, robot_coll, coll, 0.1, world_coll_weight
-        )
-        for coll in world_coll
-    ]
-
-    factors.append(self_coll_factor)
-    factors.extend(world_coll_factors)
-
-    # Solve IK.
-    joint_vars = [JointVar(joint_var_idx)]
-    graph = jaxls.FactorGraph.make(
-        factors,
-        joint_vars,
-        use_onp=False,
-    )
-    solution = graph.solve(
-        initial_vals=jaxls.VarValues.make(joint_vars),
-        trust_region=jaxls.TrustRegionConfig(lambda_initial=1.0),
-        termination=jaxls.TerminationConfig(
-            gradient_tolerance=1e-5,
-            parameter_tolerance=1e-5,
-            max_iterations=50,
-        ),
-        verbose=False,
-    )
-
-    # Update visualization.
-    joints = solution[JointVar(joint_var_idx)]
-    return joints
 
 
 if __name__ == "__main__":
