+++ src/jaxmp/extras/__init__.py
@@ -1,3 +1,4 @@
 from .urdf_loader import load_urdf as load_urdf
 from .solve_ik import solve_ik as solve_ik
+from .solve_ik import solve_ik_with_coll as solve_ik_with_coll
 from .grasp_antipodal import AntipodalGrasps as AntipodalGrasps
